refactor: #657 results/page をリスト・詳細・図に分割する
データ取得は useResultsData、表示は Status/List/Detail/Diagram に委譲し、
page を薄いオーケストレーションにする。マッチング・応募導線は不変。
